+++ src/apps/sequencer/ui/model/ProjectListModel.h
@@ -76,7 +76,8 @@ private:
         MonitorMode,
         RecordMode,
         MidiInput,
-        MidiPgmChange,
+        MidiIntegrationMode,
+        MidiProgramOffset,
         CvGateInput,
         StepsToStop,
         RecordDelay,
@@ -97,7 +98,8 @@ private:
         case MonitorMode:       return "Monitor Mode";
         case RecordMode:        return "Record Mode";
         case MidiInput:         return "MIDI Input";
-        case MidiPgmChange:     return "MIDI Pgm Chng";
+        case MidiIntegrationMode:   return "MIDI Integr.";
+        case MidiProgramOffset:     return "MIDI Pgm Off.";
         case CvGateInput:       return "CV/Gate Input";
         case StepsToStop:       return "Steps to stop";
         case RecordDelay:       return "Record Delay";
@@ -146,8 +148,11 @@ private:
         case MidiInput:
             _project.printMidiInput(str);
             break;
-        case MidiPgmChange:
-            _project.printMidiPgmChange(str);
+        case MidiIntegrationMode:
+            _project.printMidiIntegrationMode(str);
+            break;
+        case MidiProgramOffset:
+            _project.printMidiProgramOffset(str);
             break;
         case CvGateInput:
             _project.printCvGateInput(str);
@@ -200,8 +205,11 @@ private:
         case MidiInput:
             _project.editMidiInput(value, shift);
             break;
-        case MidiPgmChange:
-            _project.editMidiPgmChange(value, shift);
+       case MidiIntegrationMode:
+            _project.editMidiIntegrationMode(value, shift);
+            break;
+        case MidiProgramOffset:
+            _project.editMidiProgramOffset(value, shift);
             break;
         case CvGateInput:
             _project.editCvGateInput(value, shift);