+++ src/apps/sequencer/ui/pages/PatternPage.cpp
@@ -23,6 +23,7 @@ enum class ContextAction {
     Copy,
     Paste,
     Duplicate,
+    Save,
     Last
 };
 
@@ -371,6 +372,9 @@ void PatternPage::contextAction(int index) {
     case ContextAction::Duplicate:
         duplicatePattern();
         break;
+     case ContextAction::Save:
+        sendMidiProgramSave();
+        break;
     case ContextAction::Last:
         break;
     }
@@ -380,6 +384,8 @@ bool PatternPage::contextActionEnabled(int index) const {
     switch (ContextAction(index)) {
     case ContextAction::Paste:
         return _model.clipBoard().canPastePattern();
+    case ContextAction::Save:
+        return _engine.midiProgramChangesEnabled() && _project.midiIntegrationMalekkoEnabled();
     default:
         return true;
     }
@@ -407,5 +413,12 @@ void PatternPage::duplicatePattern() {
         _model.clipBoard().pastePattern(_project.selectedPatternIndex());
         _model.clipBoard().clear();
         showMessage("PATTERN DUPLICATED");
+    }
+}
+
+void PatternPage::sendMidiProgramSave() {
+    if (_engine.midiProgramChangesEnabled()) {
+        _engine.sendMidiProgramSave(_project.playState().trackState(0).pattern());
+        showMessage("SENT MIDI PROGRAM SAVE");
     }
 }